CR - move activitylog middleware
